Add Whitehall document types that support related links
This commit adds existing Whitehall and Specialist Publisher document types to Content Tagger that support related links. Originally, this was a list of apps that weren't able to render the sidebar (back in 2016) which was then converted to the associated document types (in 2017).

We now display the sidebar for many more document types; by allowing the related links component to be renderable when editing a content item within Content Tagger, publishers will be able to update related links