Add suggested_ordered_related_items as an editable tag type
This commit adds `suggested_ordered_related_items` as an editable tag type, allowing publishers to re-order or remove any suggested related links. New links are not able to be added to `suggested_ordered_related_items` directly, as this property is updated by an automated process and as such links may be overwritten.