Update Taxon service publishes Brexit taxon in 'cy' locale
The Brexit taxon needs to be published in the 'cy' as well as 'en' locales so
that we can serve a Welsh translation of the page. This modifies the Update
taxon Service to make a PUT content request to the Publishing API for the
Brexit taxon in the 'cy' locale.