Fix Style/MixinUsage issues
This removes top-level 'include' statements in various test files,
as well as some surprising 'module Metrics' declarations.

- Some of the 'include' statements, like 'include Taxonomy', were
solely to try and avoid 'Taxonomy::' in the tests. I've switched to
doing this, as it's just clearer.

- It's not possible to 'include' a module in a factory, so I've gone
with the suggested approach [1] of allowing the helper methods to
be called on the module, without including it.

[1]: https://github.com/thoughtbot/factory_bot/issues/564#issuecomment-24818655