Allow bulk tagging travel advice to organisation
This adds two rake tasks:
- `travel_advice:tag_edition_to_organisation[content_id,org_id]`
- `travel_advice:tag_editions_to_organisation[org_id]`

The first allows passing the `content_id` of a travel advice edition and
the `content_id` of an organisation to which the edition will be tagged.
The second allows passing the `content_id` of an organisation to which
all published travel advice editions will be tagged.

This is part of the work around the Machinery of Government changes that
will see FCO and DFID merge to become a new department. It will make it
possible to tag documents to organisations without republishing them.

Trello card: https://trello.com/c/OLlCzb7K/2096-spike-bulk-retagging-travel-advice-organisation-tagging-to-fcdo