Reference GOV.UK Docker in Technical Documentation
This uses the now somewhat standard documentation of how to run a GOV.UK
Rails app on docker.

I removed the information about general permissions as this seemed rather
dated with the concept of a tagathon participant being something of the
past and rather well documented through code.

I removed the local development permissions as these are set
automatically in seeds and not something a developer needs to do
manually.