Simplify the README
Unusually for a GOV.UK app this readme this had a lot of screenshots
which seemed rather superfluous (and a maintenance headache) and a brief
description of features.

I felt these features were sufficiently summarised in the initial
introduction so could be removed. I didn't imagine anyone gets a huge
amount of value from the screenshots (as exciting as the bubble one is)
and also removed them.