Add rake task to update travel advice related links
This updates the related links on all travel advice pages to include the
"Travel abroad: step by step" as the second item in the related links.

These updates would normally be made through the Content Tagger UI, but
there are 226 pages to update, so this would be a time consuming
process.

The rake task can be removed after it has been run.