Use RestClient to get bulk upload CSVs
RestClient follows redirects by default. It looks like the behaviour for
published CSV docs has changed recently in that the published URL redirects
to another short-lived location. Unfortunately, this breaks the workflow in
this app because Net/HTTP doesn't follow redirects automatically.

This means we end up with lots of [errors in sentry](https://sentry.io/organizations/govuk/issues/2457278761/?project=202218&query=is%3Aunresolved)
instead of tagged documents.

I've flipped some of the tests to use webmock instead of artificially producing
results from Net/Http or whatever. It means we're less tied to an implementation
and more accurately reflects how the system is behaving.