Update the remaning validators to use Errors#add
As part of the Ruby bump, using << to push objects into the
ActiveModel::Errors has been deprecated. When you make the changes
to use the #add method, the doubles start complaining as #add is not a
method on a base hash.

As each validator is only in one model, i've chosen to pass in a built/
build_stubbed instance of the model.