Fix broken taxonomy health check rake task
There is a rake task that checks the taxonomy health which does not
work. This task is run in production every day.

The task attempts to access the configuration using string keys, but
they should be symbols. This means the task fails and an error is
reported to Sentry every day.

Fixing this so the task now works and errors will no longer be reported
to Sentry.