Relax URL override regex
We were using the same regex as taxons use, which is limited to two path
segements:
- /foo
- /foo/bar

The URL override path shouldn't be limited in this way, as we should be
allowed deeper - many Whitehall URLs are several layers deep. This change
should allow:
- /foo
- /foo/bar
- /foo/bar/something/else

I indented the regexes to show how similar they are.