Fix sidekiq tests running inline
Occasionally (because of the random order of tests), other config which
sets Sidekiq to run inline leaks here. This causes the tests to fail because
the code is not run as a worker in test mode.

I've wrapped this test in a `fake!` block which ensures that this test
runs as we want and doesn't interfere with others.