Update tests to be aware of new hidden topic behaviour
The tests needs to be valid for the new behaviour in hiding
topics that are copies of mainstream browse pages. In most
cases this is simply updating what we stub, but in the case
of linkables it also adds stubbed copied topics that
should be filtered out - and will fail if they are not.