Rework test for taxonomy:untag to avoid intermittent failure
This task exits early which works oddly in Rake. Previously used `return` which is not valid (but worked by accident, because it caused a runtime error). `abort` appeared to work but fails intermittently in CI, perhaps due to test ordering? Stack Overflow recommends `next` instead. Then the test can just ensure that the untagger is never called rather than checking for a `SystemExit`.