Use Rails transactional tests instead of DatabaseCleaner gem
This allows us to lose a dependency and simplify the configuration.

We were able to replace the need to create a user before each test by
instead using the seed data - I've added Rails.application.load_seeds,
so this is automatically loaded before the suite.

This also removes the redundant Webmock.reset! which wasn't needed
before each test as Webmock does this automatically [1]

[1]: https://github.com/bblimke/webmock/blob/833291d4ce2d5927a738f929db26b594bf4fa7f5/lib/webmock/rspec.rb#L38