Remove constant redefinition
Redefining a constant produces a Ruby warning (since it's not a constant
if it varies). To resolve this I've removed the code.

As this change only applies to one use of TaxonomyQuery I understand
that if this is a problem in one place it'd be a problem in others.
Therefore my expectation is that this is not a consistent problem.

There's also a pretty high chance that this code hasn't been used
properly in years, I'm not keen to venture into that yak shave right now
though.