Port govuk_taxonomy_helpers gem into codebase
The govuk_taxonomy_helpers [1] gem is only used by this application and has
had very little development since it was created in 2017.

It seems unlikely that this tool will be used by other GOV.UK
applications, so I'm merging into this application.

Once this PR is merged I'm going to go ahead and archive the gem. This will
reduce the amount of things GOV.UK will have to maintain.

[1]: https://github.com/alphagov/govuk_taxonomy_helpers