Move TaxonBasePathStructureCheck::Taxon to individual file
This has been done to work towards resolving RSpec/FilePath linting
violations. We have a situation where TaxonBasePathStructureCheck has an
internal class of Taxon and this class has unit tests, the unit tests
for this file are in the file that tests the service.

Without trying to dig myself too big a rabbit hole here I thought the
simplest resolution would be to move the Taxon class to an individual
file and move the test to target that file.

There remains some questions about what this Taxon class actually is,
perhaps it should be a type of model?