Resolve method names in top level RSpec describes
This resolves Rubocop violations for RSpec/FilePath where it's expected
that files dedicated to testing just a single method have a dedicated
file path to specify that. Since this type of test is pretty much
non-existent in GOV.UK it seems better to adopt the standard style of
testing the class as a whole with individual describes for methods.