Make govuk_app_config update conservative
The capybara gem is currently not compatible with Puma 6 and thus this
gem update can't go ahead.

Capybara already has a fix for this which hasn't been released [1], I'm
hoping that this is a sub-dependency it'll just be updated soon and not
be of concern.

I'm hoping to avoid a pin as we don't have a direct dependency on puma
currently.

[1]: teamcapybara/capybara#2530