Update example spreadsheet for bulk tagging
The original example spreadsheet was deleted from Google Drive, so it
was 404 erroring when users clicked the link.

We've updated the link to use a new example spreadsheet. This new Google
Sheet is held in a Shared Drive, so it should persist beyond the life of
any one individual's Google account.

This was reported to us via Zendesk:
https://govuk.zendesk.com/agent/tickets/5188849