Add temporary rake task to remove legacy taxon links
Legacy taxon links are now deprecated, and we want to remove the last traces of them before removing them from the taxon schema.

This rake task takes a list of the ~170 content ids of taxons that were mapped to specialist topics as per https://docs.google.com/spreadsheets/d/1pqge1MfikjMbSoZXj74ynXllP-HIyx8ofoqRm6g0vAM/edit\#gid\=0, and for each of them makes a call to remove legacy_taxons from their links if it is present.

This can be removed once it's been confirmed that there are no remaining taxons with legacy_taxons links.