Mount /component-guide
This should be enabled on every app which pulls in govuk_publishing_components.

Similarly, moves the (deprecated) govuk_admin_template `/style-guide` route out of the dev block so that it's available on production. It was decided in https://trello.com/c/ITcz5b1G/3417-enable-component-guide-on-non-dev-environments that there's no reason we need to lock this down to dev env only.