Revert "Update the bundler version in Gemfile.lock"
That commit was an attempt to fix the build, but the underlying issue
was caused by bundler 4 being installed on the base Docker image. That
change has now been undone, with bundler pinned to an earlier version.
Since we don't know the consequences of updating bundler, this reverts
the change until upgrading bundler can be looked at properly.

This reverts commit 76bcae60e3ce9aa82ac85ad89788e505fbb53041.